Added StoryExperience UI, polishing
